refactor: move NewsSection to the components folder
