+++ src/components/NewsSection/new-big-card/__icon/new-big-card__icon.css
