+++ src/components/NewsSection/new-big-card/__row/new-big-card__row.css
