+++ src/components/NewsSection/new-big-card/__text/new-big-card__text.css
