+++ src/components/NewsSection/new-big-card/new-big-card.css
