+++ src/components/NewsSection/news-section/__column/news-section__column.css
