+++ src/components/NewsSection/news-section/__item/news-section__item.css
