+++ src/components/NewsSection/news-section/news-section.css
